fix: cleanup more types
