--- packages/shared/lib/models/Flow.ts
@@ -1,8 +1,0 @@
-export interface FlowDownloadBody {
-    id?: number;
-    name: string;
-    provider: string;
-    is_public: boolean;
-    providerConfigKey: string;
-    flowType: string;
-}
