+++ packages/shared/lib/models/Sync.ts
@@ -1,5 +1,4 @@
-import type { TimestampsAndDeleted } from './Generic.js';
-import type { ScriptTypeLiteral, Timestamps } from '@nangohq/types';
+import type { ScriptTypeLiteral, SyncResultByModel, Timestamps, TimestampsAndDeleted } from '@nangohq/types';
 
 export enum SyncStatus {
     RUNNING = 'RUNNING',
@@ -16,14 +15,6 @@ export enum SyncJobsType {
     ON_EVENT_SCRIPT = 'ON_EVENT_SCRIPT',
     ACTION = 'ACTION'
 }
-
-export interface SyncResult {
-    added: number;
-    updated: number;
-    deleted: number;
-}
-
-export type SyncResultByModel = Record<string, SyncResult>;
 
 export type SyncWithConnectionId = Sync & { connection_id: string };
 
@@ -56,22 +47,6 @@ export interface Job extends Timestamps {
     log_id?: string | null;
     result?: SyncResultByModel;
     sync_config_id?: number;
-}
-
-export interface ReportedSyncJobStatus {
-    id?: string;
-    type: SyncJobsType | 'INITIAL';
-    name?: string;
-    variant?: string;
-    connection_id?: string;
-    status: SyncStatus;
-    latestResult?: SyncResultByModel | undefined;
-    jobStatus?: SyncStatus;
-    frequency: string | null;
-    finishedAt: Date | undefined;
-    nextScheduledSyncAt: Date | null;
-    latestExecutionStatus: SyncStatus | undefined;
-    recordCount: Record<string, number>;
 }
 
 export enum SyncCommand {
