+++ packages/shared/lib/models/index.ts
@@ -2,5 +2,4 @@ export * from './Telemetry.js';
 export type * from './Generic.js';
 export type * from './Provider.js';
 export * from './Sync.js';
-export type * from './Flow.js';
 export type * from './NangoConfig.js';
