+++ packages/shared/lib/services/sync/job.service.ts
@@ -4,8 +4,8 @@ import { SyncStatus } from '../../models/Sync.js';
 import { LogActionEnum } from '../../models/Telemetry.js';
 import errorManager, { ErrorSourceEnum } from '../../utils/error.manager.js';
 
-import type { Job as SyncJob, SyncJobsType, SyncResultByModel } from '../../models/Sync.js';
-import type { ConnectionJobs } from '@nangohq/types';
+import type { Job as SyncJob, SyncJobsType } from '../../models/Sync.js';
+import type { ConnectionJobs, SyncResultByModel } from '@nangohq/types';
 
 const SYNC_JOB_TABLE = dbNamespace + 'sync_jobs';
