+++ packages/types/lib/dbConfig/db.ts
@@ -1,0 +1,4 @@
+export interface DBConfig {
+    encryption_key_hash: string | null;
+    encryption_complete: boolean;
+}
