+++ packages/types/lib/index.ts
@@ -68,6 +68,7 @@ export type * from './plans/db.js';
 export type * from './plans/http.api.js';
 export type * from './stripe/http.api.js';
 export type * from './proxy/http.api.js';
+export type * from './dbConfig/db.js';
 
 export type * from './nangoYaml/index.js';
